rename wizard pages directory
