+++ apps/chart_sync/cli.py
@@ -16,7 +16,7 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
 from apps.chart_sync.admin_api import AdminAPI
-from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
+from apps.wizard.app_pages.chart_diff.chart_diff import ChartDiffModified
 from apps.wizard.utils.env import OWIDEnv
 from etl import config
 from etl import grapher_model as gm
