+++ apps/owidbot/chart_diff.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import Session
 from structlog import get_logger
 
 from apps.chart_sync.cli import _modified_chart_ids_by_admin
-from apps.wizard.pages.chart_diff.chart_diff import ChartDiffModified
+from apps.wizard.app_pages.chart_diff.chart_diff import ChartDiffModified
 from apps.wizard.utils.env import OWID_ENV, OWIDEnv
 from etl.config import get_container_name
