+++ apps/wizard/config/config.yml
@@ -35,7 +35,7 @@ main:
       maintainer:
         - "@lucas"
         - "@daniel"
-      entrypoint: pages/expert/app.py
+      entrypoint: app_pages/expert/app.py
       emoji: "🧙"
 
 # ETL steps
@@ -81,7 +81,7 @@ etl:
     fasttrack:
       title: "Fast Track"
       alias: fasttrack
-      entrypoint: pages/fasttrack/app.py
+      entrypoint: app_pages/fasttrack/app.py
       emoji: "🏎️"
       image_url: "https://upload.wikimedia.org/wikipedia/commons/b/be/Alonso_%28Renault%29_qualifying_at_USGP_2005.jpg"
 
@@ -92,7 +92,7 @@ sections:
     apps:
       - title: "Chart Upgrader"
         alias: charts
-        entrypoint: pages/charts/__main__.py
+        entrypoint: app_pages/charts/__main__.py
         description: Prepare chart revisions
         maintainer: "@lucas"
         emoji: "🌟"
@@ -113,7 +113,7 @@ sections:
     apps:
       - title: "Dashboard"
         alias: dashboard
-        entrypoint: pages/dashboard.py
+        entrypoint: app_pages/dashboard.py
         description: Control panel for ETL steps
         maintainer: "@pablo"
         emoji: "📋"
@@ -122,21 +122,21 @@ sections:
         alias: dataexp
         description: "Explore datasets and their metadata"
         maintainer: "@lucas"
-        entrypoint: pages/dataset_explorer.py
+        entrypoint: app_pages/dataset_explorer.py
         emoji: "🔍"
         image_url: "https://upload.wikimedia.org/wikipedia/commons/c/c3/NGC_4414_%28NASA-med%29.jpg"
       - title: "Harmonizer"
         alias: harmonizer
         description: "Harmonize a column of a table"
         maintainer: "@lucas"
-        entrypoint: pages/harmonizer.py
+        entrypoint: app_pages/harmonizer.py
         emoji: "🎶"
         image_url: "https://upload.wikimedia.org/wikipedia/commons/thumb/c/c1/C_triad.svg/2560px-C_triad.svg.png"
       - title: "Insighter"
         alias: insighter
         description: "Data insights with GPT"
         maintainer: "@daniel"
-        entrypoint: pages/datainsight_robot.py
+        entrypoint: app_pages/datainsight_robot.py
         emoji: "💡"
         image_url: "https://assets.change.org/photos/7/zi/ph/ZMziPHIKGSDaDCJ-800x450-noPad.jpg?1525274743"
 
@@ -146,7 +146,7 @@ sections:
     apps:
     - title: "Indicator Upgrader"
       alias: indicator-upgrade
-      entrypoint: pages/indicator_upgrade/app.py
+      entrypoint: app_pages/indicator_upgrade/app.py
       description: Update indicator references in charts
       maintainer: "@lucas"
       emoji: "🧬"
@@ -155,7 +155,7 @@ sections:
         production: True
     - title: "Chart Diff"
       alias: chart-diff
-      entrypoint: pages/chart_diff/app.py
+      entrypoint: app_pages/chart_diff/app.py
       description: Chart differences
       maintainer: "@mojmir"
       emoji: "⚡"
@@ -171,7 +171,7 @@ sections:
       - title: "Meta Upgrader"
         description: "Migrate old metadata to new format."
         maintainer: "@veronika / @lucas"
-        entrypoint: pages/metagpt.py
+        entrypoint: app_pages/metagpt.py
         emoji: "🌟"
         alias: metagpt
         image_url: "https://cdn.pixabay.com/photo/2016/12/04/18/58/instagram-1882329_1280.png"
@@ -182,7 +182,7 @@ sections:
         alias: metaplay
         description: "How is metadata presented in data pages?"
         maintainer: "@lucas"
-        entrypoint: pages/metaplay.py
+        entrypoint: app_pages/metaplay.py
         emoji: "🏐"
         image_url: "https://upload.wikimedia.org/wikipedia/commons/d/d4/PlayEquipComboPlastic_wb.jpg"
         disable:
@@ -197,13 +197,13 @@ sections:
         alias: news
         description: "Latest ETL news"
         maintainer: "@lucas"
-        entrypoint: pages/news.py
+        entrypoint: app_pages/news.py
         emoji: "🗞️"
         image_url: "https://www.tsanet.org/wp-content/uploads/2022/08/latest-news.jpg"
       - title: "owidle"
         alias: owidle
         description: "Guess the country game!"
         maintainer: "@lucas"
-        entrypoint: pages/owidle.py
+        entrypoint: app_pages/owidle.py
         emoji: "👾"
         image_url: "https://upload.wikimedia.org/wikipedia/en/thumb/e/e0/WPVG_icon_2016.svg/160px-WPVG_icon_2016.svg.png"
