--- apps/wizard/pages/__init__.py
@@ -1,1 +1,0 @@
-"""Add here the pages of the wizard app."""
