--- apps/wizard/pages/expert/__init__.py
@@ -1,1 +1,0 @@
-"""GPT-based chat with ETL knowledge."""
