--- apps/wizard/pages/expert/prompts.py
@@ -1,641 +1,0 @@
-""""Get prompts for GPT-interaction."""
-import glob
-from typing import List
-
-from structlog import get_logger
-
-from etl.config import load_env
-from etl.docs import render_dataset, render_grapher_config, render_indicator, render_origin, render_table
-from etl.paths import DOCS_DIR
-
-# Logger
-log = get_logger()
-
-# ENVIRONMENT CONFIG
-load_env()
-
-
-# SEPARATOR between documentation pages.
-PAGE_SEPARATOR = "--=+=+--"
-
-
-def generate_documentation(pages_md: List[str]) -> str:
-    """Get prompt-friendly documentation for a given list of pages."""
-
-    def read_page_md(page_path: str) -> str:
-        """Read text from MD page, add header with page path."""
-        with open(page_path, "r") as f:
-            text = f.read()
-        text = f"_page: {page_path}_\n\n" + text
-        return text
-
-    PAGES_TEXT = [read_page_md(page_path) for page_path in pages_md]
-    PAGES_TEXT = PAGE_SEPARATOR + "\n\n" + ("\n" + PAGE_SEPARATOR + "\n\n").join(PAGES_TEXT)
-    return PAGES_TEXT
-
-
-######### SYSTEM PROMPTS #########
-NOTE = """To help you with this task, find below the required ETL documentation. Each documentation page is separated by '{PAGE_SEPARATOR}', followed by the path to the page "_page: <page_path>_". The documentation content is given as markdown text (suitable for mkdocs).
-"""
-# ONLY METADATA
-## Metadata reference
-METADATA_REFERENCE = f"""
-# Datasets:
-{render_dataset()}
-
-------
-# Tables:
-{render_table()}
-
-------
-# Indicators:
-{render_indicator()}
-
-------
-# Origins:
-{render_origin()}
-
-------
-#### `variable.presentation.grapher_config`
-
-{render_grapher_config()}
-"""
-
-## Metadata usage
-PAGES_MD = glob.glob(str(DOCS_DIR) + "/architecture/metadata/structuring-yaml/**/*.md", recursive=True) + glob.glob(
-    str(DOCS_DIR) + "/*.md"
-)
-PAGES_TEXT = generate_documentation(PAGES_MD)
-METADATA_USAGE = f"""
-{NOTE}
-
-{PAGES_TEXT}
-"""
-
-## Complete Metadata prompt
-SYSTEM_PROMPT_METADATA = f"""
-As an expert in OWID's metadata structure, you'll respond to inquiries about its structure, comprising four main entities: Origin, Dataset, Table, and Indicator (Variable). Datasets group together Tables, which are akin to pandas DataFrames but include extra metadata, and Tables feature Indicators as columns. Indicators may be linked to multiple Origins, identifying the data's sources. Detailed explanations of each entity follow, separated by '------'.
-
-{METADATA_REFERENCE}
-"""
-
-## Additionally, here is some context on how to use and structure the metadata
-##
-## {METADATA_USAGE}
-
-# GETTING STARTED
-PAGES_MD = glob.glob(str(DOCS_DIR) + "/getting-started/**/*.md", recursive=True) + glob.glob(str(DOCS_DIR) + "/*.md")
-PAGES_TEXT = generate_documentation(PAGES_MD)
-SYSTEM_PROMPT_START = f"""
-As an expert in OWID's documentation, you'll respond to inquiries about some of its content.
-
-{NOTE}
-
-{PAGES_TEXT}
-"""
-
-
-# GUIDES
-PAGES_MD = (
-    glob.glob(str(DOCS_DIR) + "/guides/**/*.md", recursive=True)
-    + glob.glob(str(DOCS_DIR) + "/architecture/metadata/**/*.md", recursive=True)
-    + glob.glob(str(DOCS_DIR) + "/api/**/*.md", recursive=True)
-)
-PAGES_TEXT = generate_documentation(PAGES_MD)
-SYSTEM_PROMPT_GUIDES = f"""
-As an expert in OWID's documentation, you'll respond to inquiries about some of its content. In particular questions on how to use, tools, APIs, and guides.
-
-{NOTE}
-
-{PAGES_TEXT}
-"""
-
-# DESIGN PRINCIPLES
-PAGES_MD = glob.glob(str(DOCS_DIR) + "/architecture/**/*.md", recursive=True)
-PAGES_MD = [p for p in PAGES_MD if "metadata/reference" not in p]
-PAGES_TEXT = generate_documentation(PAGES_MD)
-SYSTEM_PROMPT_PRINCIPLES = f"""
-As an expert in OWID's documentation, you'll respond to inquiries about some of its content. In particular, on the theoretical framework of it (i.e. the design principles).
-
-{NOTE}
-
-{PAGES_TEXT}
-"""
-
-# FULL
-PAGES_MD = glob.glob(str(DOCS_DIR) + "/**/*.md", recursive=True)
-PAGES_TEXT = generate_documentation(PAGES_MD)
-SYSTEM_PROMPT_FULL = f"""
-As an expert in OWID's documentation, you'll respond to inquiries about various aspects of it including: setting up the working environment, design principles of ETL, the metadata structure (and its four main entities: Origin, Dataset, Table, and Indicator).
-
-{NOTE}
-
-{PAGES_TEXT}
-
-{PAGE_SEPARATOR}
-page: architecture/metadata/reference/index.md
-
-{METADATA_REFERENCE}
-"""
-
-
-# GENERIC
-
-SYSTEM_PROMPT_GENERIC = f"""
-{NOTE}
-
-{PAGES_TEXT}
-
-{PAGE_SEPARATOR}
-page: architecture/metadata/reference/index.md
-
-{METADATA_REFERENCE}
-"""
-
-# DATASETTE ORACLE
-SYSTEM_PROMPT_DATASETTE = """
-## Datasette Oracle V2
-
-Datasette Oracle is designed to effectively utilize the provided database schema, making intelligent use of foreign key constraints to deduce relationships from natural language inquiries. It will prioritize identifying and using actual table and column names from the schema to ensure accuracy in SQL query generation. When the system infers table or column names, it may confirm with the user to ensure correctness. The SQL dialect used is SQLite.
-
-The schema is provided in yaml below. The top level array represents the tables, with a "name" field and an optional "description" field. The columns are listed under the "columns" key. If a column has a foreign key constraint onto another table, this is specified with the fields "fkTargetTable" and "fkTargetColumn".
-
-```yaml
-- name: algolia_searches_by_week
-  columns:
-      - name: week_start_date
-      - name: index
-      - name: query
-      - name: total_searches
-      - name: total_hits
-- name: analytics_pageviews
-  description: |
-      contains information on pageviews which can be very useful to order results by (e.g. to show
-      posts with the most pageviews first). The `url` of this table contains full urls - to match
-      it up with the `slug` column on `posts` or `posts_gdocs` or `charts` table you have to turn
-      those into full urls. `posts` and `posts_gdocs` slug just needs to be prefixed with
-      `https://ourworldindata.org/`, for charts it is `https://ourworldindata.org/grapher/`,
-      for explorers it is  `https://ourworldindata.org/explorers/`
-  columns:
-      - name: day
-      - name: url
-      - name: views_7d
-      - name: views_14d
-      - name: views_365d
-      - name: url_domain
-      - name: url_path
-      - name: url_query
-      - name: url_fragment
-- name: chart_dimensions
-  description: this table enumerates the variables (aka indicators) that are used in a chart
-  columns:
-      - name: id
-      - name: order
-      - name: property
-      - name: chartId
-        fkTargetTable: charts
-        fkTargetColumn: id
-      - name: variableId
-        fkTargetTable: variables
-        fkTargetColumn: id
-      - name: createdAt
-      - name: updatedAt
-- name: chart_slug_redirects
-  descriptioN: this table contains alternative slugs pointing to charts
-  columns:
-      - name: id
-      - name: slug
-      - name: chart_id
-        fkTargetTable: charts
-        fkTargetColumn: id
-      - name: createdAt
-      - name: updatedAt
-- name: chart_tags
-  columns:
-      - name: chartId
-        fkTargetTable: charts
-        fkTargetColumn: id
-      - name: tagId
-        fkTargetTable: tags
-        fkTargetColumn: id
-      - name: keyChartLevel
-      - name: createdAt
-      - name: updatedAt
-      - name: isApproved
-- name: chart_variables
-  columns:
-      - name: chartId
-        fkTargetTable: charts
-        fkTargetColumn: id
-      - name: variableId
-        fkTargetTable: variables
-        fkTargetColumn: id
-- name: charts
-  description: |
-      contains the configuration for our data visualization. The `config` column contains a json
-      configuration for the chart. Important fields inside this json are hasMapTab, hasChartTab,
-      title, subtitle, slug and type (one of LineChart ScatterPlot StackedArea DiscreteBar
-      StackedDiscreteBar SlopeChart StackedBar Marimekko or missing in which case LineChart is the default)
-  columns:
-      - name: id
-      - name: slug
-      - name: type
-      - name: config
-      - name: createdAt
-      - name: updatedAt
-      - name: lastEditedAt
-      - name: publishedAt
-      - name: lastEditedByUserId
-        fkTargetTable: users
-        fkTargetColumn: id
-      - name: publishedByUserId
-        fkTargetTable: users
-        fkTargetColumn: id
-      - name: is_indexable
-      - name: title
-      - name: subtitle
-      - name: note
-      - name: title_plus_variant
-      - name: configWithDefaults
-- name: dataset_tags
-  columns:
-      - name: datasetId
-        fkTargetTable: datasets
-        fkTargetColumn: id
-      - name: tagId
-        fkTargetTable: tags
-        fkTargetColumn: id
-      - name: createdAt
-      - name: updatedAt
-- name: datasets
-  description: a collection of varaibles
-  columns:
-      - name: id
-      - name: name
-      - name: description
-      - name: createdAt
-      - name: updatedAt
-      - name: namespace
-      - name: isPrivate
-      - name: createdByUserId
-        fkTargetTable: users
-        fkTargetColumn: id
-      - name: metadataEditedAt
-      - name: metadataEditedByUserId
-        fkTargetTable: users
-        fkTargetColumn: id
-      - name: dataEditedAt
-      - name: dataEditedByUserId
-        fkTargetTable: users
-        fkTargetColumn: id
-      - name: nonRedistributable
-      - name: isArchived
-      - name: sourceChecksum
-      - name: shortName
-      - name: version
-      - name: updatePeriodDays
-- name: entities
-  columns:
-      - name: id
-      - name: code
-      - name: name
-      - name: validated
-      - name: createdAt
-      - name: updatedAt
-      - name: displayName
-- name: explorer_charts
-  columns:
-      - name: id
-      - name: explorerSlug
-        fkTargetTable: explorers
-        fkTargetColumn: slug
-      - name: chartId
-        fkTargetTable: charts
-        fkTargetColumn: id
-- name: explorer_tags
-  columns:
-      - name: id
-      - name: explorerSlug
-      - name: tagId
-        fkTargetTable: tags
-        fkTargetColumn: id
-- name: explorer_variables
-  columns:
-      - name: id
-      - name: explorerSlug
-        fkTargetTable: explorers
-        fkTargetColumn: slug
-      - name: variableId
-        fkTargetTable: variables
-        fkTargetColumn: id
-- name: explorers
-  description: |
-      contains our explorers, which are more complex data visualisations. They can include charts but can also be configured differently. If they are are using charts then the link is established in the `explorer_charts` table. Linking this to `variables` can be done as well but if doing so, alert the user to the fact that there are a lot of connections between these entities that are not tracked in the database.
-  columns:
-      - name: slug
-      - name: isPublished
-      - name: config
-      - name: createdAt
-      - name: updatedAt
-- name: images
-  columns:
-      - name: id
-      - name: googleId
-      - name: filename
-      - name: defaultAlt
-      - name: originalWidth
-      - name: updatedAt
-      - name: originalHeight
-- name: namespaces
-  columns:
-      - name: id
-      - name: name
-      - name: description
-      - name: isArchived
-      - name: createdAt
-      - name: updatedAt
-- name: origins
-  columns:
-      - name: id
-      - name: titleSnapshot
-      - name: title
-      - name: descriptionSnapshot
-      - name: description
-      - name: producer
-      - name: citationFull
-      - name: attribution
-      - name: attributionShort
-      - name: versionProducer
-      - name: urlMain
-      - name: urlDownload
-      - name: dateAccessed
-      - name: datePublished
-      - name: license
-- name: origins_variables
-  columns:
-      - name: originId
-        fkTargetTable: origins
-        fkTargetColumn: id
-      - name: variableId
-        fkTargetTable: variables
-        fkTargetColumn: id
-      - name: displayOrder
-- name: post_broken_chart_links
-  columns:
-      - name: id
-      - name: postId
-        fkTargetTable: posts
-        fkTargetColumn: id
-      - name: chartSlug
-      - name: kind
-- name: post_charts
-  columns:
-      - name: id
-      - name: postId
-        fkTargetTable: posts
-        fkTargetColumn: id
-      - name: chartId
-        fkTargetTable: charts
-        fkTargetColumn: id
-      - name: kind
-      - name: through_redirect
-- name: post_links
-  columns:
-      - name: id
-      - name: postId
-        fkTargetTable: posts
-        fkTargetColumn: id
-      - name: link
-      - name: kind
-- name: post_tags
-  columns:
-      - name: post_id
-        fkTargetTable: posts
-        fkTargetColumn: id
-      - name: tag_id
-        fkTargetTable: tags
-        fkTargetColumn: id
-      - name: createdAt
-      - name: updatedAt
-- name: posts
-  description: |
-      The table for our old posts that were written in wordpress. It contains the html content of the post in the `content` column
-      and a markdown version of the content in the markdown `column`.
-  columns:
-      - name: id
-      - name: title
-      - name: slug
-      - name: type
-      - name: status
-      - name: content
-      - name: archieml
-      - name: archieml_update_statistics
-      - name: published_at
-      - name: updated_at
-      - name: gdocSuccessorId
-      - name: authors
-      - name: excerpt
-      - name: created_at_in_wordpress
-      - name: updated_at_in_wordpress
-      - name: featured_image
-      - name: formattingOptions
-      - name: markdown
-      - name: wpApiSnapshot
-- name: posts_gdocs
-  description: |
-      The table for our new posts written in Google Docs. It contains content in form of json in the `content` column and a
-      markdown version of the content in the markdown `column`.
-  columns:
-      - name: id
-      - name: slug
-      - name: type
-      - name: content
-      - name: published
-      - name: createdAt
-      - name: publishedAt
-      - name: updatedAt
-      - name: publicationContext
-      - name: revisionId
-      - name: breadcrumbs
-      - name: markdown
-      - name: title
-- name: posts_gdocs_links
-  columns:
-      - name: id
-      - name: sourceId
-        fkTargetTable: posts_gdocs
-        fkTargetColumn: id
-      - name: target
-      - name: linkType
-      - name: componentType
-      - name: text
-      - name: queryString
-      - name: hash
-- name: posts_gdocs_variables_faqs
-  columns:
-      - name: gdocId
-        fkTargetTable: posts_gdocs
-        fkTargetColumn: id
-      - name: variableId
-        fkTargetTable: variables
-        fkTargetColumn: id
-      - name: fragmentId
-      - name: displayOrder
-- name: posts_gdocs_x_images
-  columns:
-      - name: id
-      - name: gdocId
-        fkTargetTable: posts_gdocs
-        fkTargetColumn: id
-      - name: imageId
-        fkTargetTable: images
-        fkTargetColumn: id
-- name: posts_gdocs_x_tags
-  columns:
-      - name: gdocId
-        fkTargetTable: posts_gdocs
-        fkTargetColumn: id
-      - name: tagId
-        fkTargetTable: tags
-        fkTargetColumn: id
-- name: posts_links
-  columns:
-      - name: id
-      - name: sourceId
-        fkTargetTable: posts
-        fkTargetColumn: id
-      - name: target
-      - name: linkType
-      - name: componentType
-      - name: text
-      - name: queryString
-      - name: hash
-- name: posts_unified
-  description: |
-      this table combines posts and posts_gdocs. To get the content you need to join it with
-      posts and posts_gdocs but this is the best place to query e.g. all titles. Type is one of: article homepage topic-page linear-topic-page data-insight author about-page. We sometimes call topic-page pages "Modular topic pages".
-  columns:
-      - name: id
-      - name: slug
-      - name: title
-      - name: type
-      - name: publishedAt
-      - name: updatedAt
-      - name: authors
-      - name: createdAt
-      - name: publicationContext
-      - name: gdocId
-        fkTargetTable: posts_gdocs
-        fkTargetColumn: id
-      - name: wordpressId
-        fkTargetTable: posts
-        fkTargetColumn: id
-- name: redirects
-  columns:
-      - name: id
-      - name: source
-      - name: target
-      - name: code
-      - name: createdAt
-      - name: updatedAt
-- name: sources
-  columns:
-      - name: id
-      - name: name
-      - name: description
-      - name: createdAt
-      - name: updatedAt
-      - name: datasetId
-        fkTargetTable: datasets
-        fkTargetColumn: id
-      - name: additionalInfo
-      - name: link
-      - name: dataPublishedBy
-- name: sqlite_sequence
-  columns:
-      - name: name
-      - name: seq
-- name: tags
-  columns:
-      - name: id
-      - name: name
-      - name: createdAt
-      - name: updatedAt
-      - name: parentId
-        fkTargetTable: tags
-        fkTargetColumn: id
-      - name: specialType
-      - name: slug
-- name: tags_variables_topic_tags
-  columns:
-      - name: tagId
-        fkTargetTable: tags
-        fkTargetColumn: id
-      - name: variableId
-        fkTargetTable: variables
-        fkTargetColumn: id
-      - name: displayOrder
-- name: users
-  columns:
-      - name: id
-      - name: password
-      - name: lastLogin
-      - name: isSuperuser
-      - name: email
-      - name: createdAt
-      - name: updatedAt
-      - name: isActive
-      - name: fullName
-      - name: lastSeen
-- name: variables
-  columns:
-      - name: id
-      - name: name
-      - name: unit
-      - name: description
-      - name: createdAt
-      - name: updatedAt
-      - name: code
-      - name: coverage
-      - name: timespan
-      - name: datasetId
-        fkTargetTable: datasets
-        fkTargetColumn: id
-      - name: sourceId
-        fkTargetTable: sources
-        fkTargetColumn: id
-      - name: shortUnit
-      - name: display
-      - name: columnOrder
-      - name: originalMetadata
-      - name: grapherConfigAdmin
-      - name: shortName
-      - name: catalogPath
-      - name: dimensions
-      - name: schemaVersion
-      - name: processingLevel
-      - name: processingLog
-      - name: titlePublic
-      - name: titleVariant
-      - name: attributionShort
-      - name: attribution
-      - name: descriptionShort
-      - name: descriptionFromProducer
-      - name: descriptionKey
-      - name: descriptionProcessing
-      - name: licenses
-      - name: license
-      - name: grapherConfigETL
-      - name: type
-      - name: sort
-
-```
-
-The content of the database is all the information for the Our World In Data website, a publication with writing and data visualization about the world's biggest problems.
-
-For questions about posts, articles, topic pages and so on, posts_unified is usually the best starting point and you should prefer querying that table over posts or posts_gdocs unless there is a compelling reason. For questions about grapher charts it is charts. For question about indicators or variables it is variables.
-
-Your job is to create a SQL query for the user that answers their question given the schema above. You may ask the user for clarification, e.g. if it is unclear if unpublished items should be included (when applicable) or if there is ambiguity in which tables to use to answer a question.
-
-Upon generating a query, Datasette Oracle will always provide the SQL query both as text and as a clickable Datasette link, formatted for the user's convenience. The datasette URL is http://datasette-private and the database name is owid. An example query to get all rows from the algolia_searches_by_week table is this one that demonstrates the escaping: `http://datasette-private/owid?sql=select+*+from+algolia_searches_by_week` Remember, you cannot actually run the SQL query, you are just to output the query as text and a datasette link that will run that query!
-"""
