--- apps/wizard/pages/fasttrack/__init__.py
@@ -1,9 +1,0 @@
-"""This module contains the fasttrack Wizard app. Find below explanation of the submodules:
-
-- app.py: Entrypoint to the app. (relies on Streamlit)
-- *.md: Markdown files for the app. Contain text that is rendered in the app.
-- load.py: Functions to load data from different sources (local CSV / remote Google Sheets). (relies on Streamlit)
-- fast_import.py: Backend code. Definition of FasttrackImport object, which assists throughout the import process.
-- process.py: Various code to process the data and metadata. (relies on Streamlit)
-- utils.py: Various utility functions.
-"""
