--- apps/wizard/pages/fasttrack/cookiecutter/{{cookiecutter.namespace}}/{{cookiecutter.version}}/{{cookiecutter.short_name}}.py
@@ -1,22 +1,0 @@
-from etl.helpers import PathFinder, create_dataset, get_metadata_path
-from etl.snapshot import Snapshot
-
-paths = PathFinder(__file__)
-
-
-def run(dest_dir: str) -> None:
-    # load snapshot
-    snap = Snapshot("{{cookiecutter.namespace}}/{{cookiecutter.version}}/{{cookiecutter.short_name}}.csv")
-
-    # load data
-    tb = snap.read_csv()
-
-    # add table, update metadata from *.meta.yml and save
-    ds = create_dataset(dest_dir, tables=[tb.set_index(["country", "year"])], default_metadata=snap.metadata)
-
-    # override metadata if necessary
-    meta_path = get_metadata_path(dest_dir).with_suffix(".override.yml")
-    if meta_path.exists():
-        ds.update_metadata(meta_path)
-
-    ds.save()
