--- apps/wizard/pages/fasttrack/markdown/fasttrack_gsheet_update.md
@@ -1,3 +1,0 @@
-#### Updating from an existing Google Sheet
-
-If you have previously imported a file from Google Sheets, this file is already being tracked by the system. Therefore, the dataset should appear in the list of datasets in the dropdown menu below.
