+++ apps/wizard/scripts/news.py
@@ -4,7 +4,7 @@ from typing import Tuple
 import requests
 from structlog import get_logger
 
-from apps.wizard.pages.expert.prompts import SYSTEM_PROMPT_GENERIC
+from apps.wizard.app_pages.expert.prompts import SYSTEM_PROMPT_GENERIC
 from apps.wizard.utils.db import WizardDB
 from apps.wizard.utils.gpt import GPTQuery, OpenAIWrapper, get_cost_and_tokens
