+++ tests/apps/wizard/pages/expert/test_prompts.py
@@ -1,2 +1,2 @@
 def test_load_module():
-    from apps.wizard.pages.expert import prompts  # noqa: F401
+    from apps.wizard.app_pages.expert import prompts  # noqa: F401
